Final spimbot.s Contest Submission
This was the implementation of the bot as of two hours before the competition. I was unhappy with the implementation of the bunny searching algorithm, so I continued to work on that while having this as what was to be submitted. When I failed to make something better in time, this was the bot that competed in the tournament. This bot landed us 5th place of something like 70 teams.